+++ server/routes/category/controller.js
@@ -1,3 +1,4 @@
+import { isValidOId } from '../../helpers/types'
 import Category from '../../models/category'
 
 export const getAllCategories = (req, res) =>
@@ -24,19 +25,22 @@ export const getAllInactiveCategories = (req, res) =>
 
     })
 
-export const getCategoryById = (req, res) =>
+export const getCategoryById = (req, res) => {
+
+  if(!isValidOId(req.params.id)) return res.boom.badRequest('Invalid Id')
 
   Category.findOne(
     { _id: req.params.id, isActive: true },
     { isActive: 0, oldId: 0 }
   )
-    .exec(async (err, category) => {
+    .exec(async (err, doc) => {
 
       if (err) return res.boom.badImplementation('', { error: err })
-      if (!category) return res.boom.notFound('Category not found')
-      return res.status(200).send({ doc: category })
+      if (!doc || !doc._id) return res.boom.notFound('Category not found')
+      return res.status(200).send({ doc: doc })
 
     })
+}
 
 export const saveCategory = (req, res) => {
 
@@ -65,6 +69,8 @@ export const saveCategory = (req, res) => {
 
 export const updateCategory = (req, res) => {
 
+  if(!isValidOId(req.params.id)) return res.boom.badRequest('Invalid Id')
+
   Category.findOneAndUpdate(
     { isActive: true, _id: req.params.id },
     {
@@ -83,7 +89,7 @@ export const updateCategory = (req, res) => {
       }
 
       if (err) return res.boom.badImplementation('', { error: err })
-      if (!doc) return res.boom.notFound('Category not found')
+      if (!doc || !doc._id) return res.boom.notFound('Category not found')
       return res.status(200).send({ message: 'Category updated!', doc: refreshCat })
 
     }
@@ -93,6 +99,8 @@ export const updateCategory = (req, res) => {
 
 export const activeCategory = (req, res) => {
 
+  if(!isValidOId(req.params.id)) return res.boom.badRequest('Invalid Id')
+
   Category.findOneAndUpdate(
     { isActive: false, _id: req.params.id },
     { isActive: true },
@@ -100,7 +108,7 @@ export const activeCategory = (req, res) => {
     (err, doc) => {
 
       if (err) return res.boom.badImplementation('', { error: err })
-      if (!doc) return res.boom.notFound('Category not found')
+      if (!doc || !doc._id) return res.boom.notFound('Category not found')
       return res.status(200).send({ message: 'Category active!', doc })
 
     }
@@ -109,6 +117,8 @@ export const activeCategory = (req, res) => {
 }
 
 export const deleteCategory = (req, res) => {
+
+  if(!isValidOId(req.params.id)) return res.boom.badRequest('Invalid Id')
 
   Category.findOneAndUpdate(
     { _id: req.params.id },
@@ -125,12 +135,14 @@ export const deleteCategory = (req, res) => {
 
 export const deleteCategoryDeep = (req, res) => {
 
+  if(!isValidOId(req.params.id)) return res.boom.badRequest('Invalid Id')
+
   Category.findByIdAndRemove(
     req.params.id,
     (err, doc) => {
 
       if (err) return res.boom.badImplementation('', { error: err })
-      if (!doc) return res.boom.notFound('Category not found')
+      if (!doc || !doc._id) return res.boom.notFound('Category not found')
       return res.status(200).send({ message: 'Category deep removed!' })
 
     }
